refactor utils->shared
