+++ shared/response_mode_handler.go
@@ -1,4 +1,4 @@
-package utils
+package shared
 
 import (
 	"errors"
