+++ shared/signature_verification_key.go
@@ -1,4 +1,4 @@
-package utils
+package shared
 
 import (
 	"context"
